# optimise getUnpaddedString with SWAR padding search