--- a/pinot-segment-local/src/main/java/org/apache/pinot/segment/local/io/util/FixedByteValueReaderWriter.java
+++ b/pinot-segment-local/src/main/java/org/apache/pinot/segment/local/io/util/FixedByteValueReaderWriter.java
@@ -57,34 +57,40 @@ public final class FixedByteValueReaderWriter implements ValueReader {
   public String getUnpaddedString(int index, int numBytesPerValue, byte paddingByte, byte[] buffer) {
     assert buffer.length >= numBytesPerValue;
     long startOffset = (long) index * numBytesPerValue;
-    int written = 0;
     long pattern = (paddingByte & 0xFFL) * 0x101010101010101L;
-    boolean le = _dataBuffer.order() == ByteOrder.LITTLE_ENDIAN;
     ByteBuffer wrapper = ByteBuffer.wrap(buffer);
-    if (le) {
+    if (_dataBuffer.order() == ByteOrder.LITTLE_ENDIAN) {
       wrapper.order(ByteOrder.LITTLE_ENDIAN);
     }
+    int position = 0;
     for (int i = 0; i < ((numBytesPerValue >>> 3) << 3); i += 8) {
       long word = _dataBuffer.getLong(startOffset + i);
       wrapper.putLong(i, word);
       long zeroed = word ^ pattern;
       long tmp = (zeroed & 0x7F7F7F7F7F7F7F7FL) + 0x7F7F7F7F7F7F7F7FL;
       tmp = ~(tmp | zeroed | 0x7F7F7F7F7F7F7F7FL);
-      int end = le
-          ? Long.numberOfTrailingZeros(tmp) >>> 3
-          : Long.numberOfLeadingZeros(tmp) >>> 3;
-      written += end;
-      if (end < 8) {
-        return new String(buffer, 0, written, UTF_8);
+      if (tmp == 0) {
+       position += 8;
+      } else {
+        position += _dataBuffer.order() == ByteOrder.LITTLE_ENDIAN
+            ? Long.numberOfTrailingZeros(tmp) >>> 3
+            : Long.numberOfLeadingZeros(tmp) >>> 3;
+        return new String(buffer, 0, position, UTF_8);
       }
     }
-    for (; written < numBytesPerValue; written++) {
-      buffer[written] = _dataBuffer.getByte(startOffset + written);
-      if (buffer[written] == paddingByte) {
+    return getUnpaddedStringTail(startOffset, position, numBytesPerValue, paddingByte, buffer);
+  }
+
+  private String getUnpaddedStringTail(long startOffset, int position, int numBytesPerValue, byte paddingByte,
+      byte[] buffer) {
+    for (; position < numBytesPerValue; position++) {
+      byte b = _dataBuffer.getByte(startOffset + position);
+      if (b == paddingByte) {
         break;
       }
+      buffer[position] = b;
     }
-    return new String(buffer, 0, written, UTF_8);
+    return new String(buffer, 0, position, UTF_8);
   }
 
   @Override