## Description
This method was the hottest method on the pinot server in a string-heavy load test:

<img width="936" alt="Screenshot 2021-11-05 at 12 33 13" src="https://user-images.githubusercontent.com/16439049/140510773-506129fb-bbb2-4fd2-8435-4063b3030167.png">

This change performs the search for padding bytes 8 bytes at a time which accelerates the search.

Before
```
Benchmark                                        (_length)  (_nativeOrder)  (_paddingByte)  (_values)  Mode  Cnt       Score       Error  Units
BenchmarkFixedByteValueReaderWriter.readStrings          8            true              42       4096  avgt    5     144.520 ±     3.255  us/op
BenchmarkFixedByteValueReaderWriter.readStrings          8           false              42       4096  avgt    5     157.459 ±    43.499  us/op
BenchmarkFixedByteValueReaderWriter.readStrings         32            true              42       4096  avgt    5     274.456 ±     3.420  us/op
BenchmarkFixedByteValueReaderWriter.readStrings         32           false              42       4096  avgt    5     277.324 ±    11.429  us/op
BenchmarkFixedByteValueReaderWriter.readStrings       1024            true              42       4096  avgt    5    4987.751 ±   204.235  us/op
BenchmarkFixedByteValueReaderWriter.readStrings       1024           false              42       4096  avgt    5    4978.517 ±   306.599  us/op
BenchmarkFixedByteValueReaderWriter.readStrings      65536            true              42       4096  avgt    5  361066.397 ± 22635.672  us/op
BenchmarkFixedByteValueReaderWriter.readStrings      65536           false              42       4096  avgt    5  330063.036 ± 22979.931  us/op
```

After
```
Benchmark                                        (_length)  (_nativeOrder)  (_paddingByte)  (_values)  Mode  Cnt      Score      Error  Units
BenchmarkFixedByteValueReaderWriter.readStrings          8            true              42       4096  avgt    5    127.794 ±    2.571  us/op
BenchmarkFixedByteValueReaderWriter.readStrings          8           false              42       4096  avgt    5    130.723 ±   10.582  us/op
BenchmarkFixedByteValueReaderWriter.readStrings         32            true              42       4096  avgt    5    192.249 ±    3.020  us/op
BenchmarkFixedByteValueReaderWriter.readStrings         32           false              42       4096  avgt    5    192.169 ±   15.036  us/op
BenchmarkFixedByteValueReaderWriter.readStrings       1024            true              42       4096  avgt    5    982.148 ±   52.506  us/op
BenchmarkFixedByteValueReaderWriter.readStrings       1024           false              42       4096  avgt    5   1025.949 ±   59.918  us/op
BenchmarkFixedByteValueReaderWriter.readStrings      65536            true              42       4096  avgt    5  62129.454 ± 1744.739  us/op
BenchmarkFixedByteValueReaderWriter.readStrings      65536           false              42       4096  avgt    5  65727.304 ± 1073.443  us/op
```

## Upgrade Notes
Does this PR prevent a zero down-time upgrade? (Assume upgrade order: Controller, Broker, Server, Minion)
* [ ] Yes (Please label as **<code>backward-incompat</code>**, and complete the section below on Release Notes)

Does this PR fix a zero-downtime upgrade introduced earlier?
* [ ] Yes (Please label this as **<code>backward-incompat</code>**, and complete the section below on Release Notes)

Does this PR otherwise need attention when creating release notes? Things to consider:
- New configuration options
- Deprecation of configurations
- Signature changes to public methods/interfaces
- New plugins added or old plugins removed
* [ ] Yes (Please label this PR as **<code>release-notes</code>** and complete the section on Release Notes)
## Release Notes
<!-- If you have tagged this as either backward-incompat or release-notes,
you MUST add text here that you would like to see appear in release notes of the
next release. -->

<!-- If you have a series of commits adding or enabling a feature, then
add this section only in final commit that marks the feature completed.
Refer to earlier release notes to see examples of text.
-->
## Documentation
<!-- If you have introduced a new feature or configuration, please add it to the documentation as well.
See https://docs.pinot.apache.org/developers/developers-and-contributors/update-document
-->