@Jackie-Jiang 2021-11-05
> Don't merge this, I don't think it has any test coverage beyond quick start (which is a bit concerning)

All the tests using `StringDictionary` should cover this part. But yes we don't have a dedicate test for `FixedByteValueReaderWriter`