@richardstartin 2021-11-05
I just reran the benchmark after the change and got similar results.