@richardstartin 2021-11-06
> Very smart algorithm!
This algorithm is described here: http://graphics.stanford.edu/~seander/bithacks.html#ZeroInWord

For the record, the technique comes from chapter 6 of Hackers Delight. I will add an attribution comment, but it’s well known to the point of being ubiquitous.